subscriber/requirements.txt
google-cloud-pubsub
requests

